+++ lib/webpacker.rb
@@ -1,24 +1,28 @@
+require "active_support/core_ext/module/attribute_accessors"
+require "active_support/logger"
+require "active_support/tagged_logging"
+
 module Webpacker
   extend self
 
-  def bootstrap
-    Webpacker::Env.load
-    Webpacker::Configuration.load
-    Webpacker::Manifest.load
+  def instance=(instance)
+    @instance = instance
   end
 
-  def compile
-    Webpacker::Compiler.compile
-    Webpacker::Manifest.load
+  def instance
+    @instance ||= Webpacker::Instance.new
   end
 
-  def env
-    Webpacker::Env.current.inquiry
-  end
+  delegate :logger, :logger=, :env, to: :instance
+  delegate :config, :compiler, :manifest, :commands, :dev_server, to: :instance
+  delegate :bootstrap, :clobber, :compile, to: :commands
 end
 
-require "webpacker/env"
+require "webpacker/instance"
 require "webpacker/configuration"
 require "webpacker/manifest"
 require "webpacker/compiler"
+require "webpacker/commands"
+require "webpacker/dev_server"
+
 require "webpacker/railtie" if defined?(Rails)