+++ test/compiler_test.rb
@@ -1,25 +1,20 @@
-require "webpacker_test"
+require "webpacker_test_helper"
 
 class CompilerTest < Minitest::Test
-  def test_default_watched_paths
-    assert_equal Webpacker::Compiler.default_watched_paths, ["app/javascript/**/*", "yarn.lock", "package.json", "config/webpack/**/*"]
-  end
-
-  def test_empty_watched_paths
-    assert_equal Webpacker::Compiler.watched_paths, []
-  end
-
-  def test_watched_paths
-    Webpacker::Compiler.stub :watched_paths, ["Gemfile"] do
-      assert_equal Webpacker::Compiler.watched_paths, ["Gemfile"]
+  def setup
+    Webpacker.compiler.send(:compilation_digest_path).tap do |path|
+      path.delete if path.exist?
     end
   end
 
-  def test_cache_dir
-    assert_equal Webpacker::Compiler.cache_dir, "tmp/webpacker"
+  def test_custom_environment_variables
+    assert Webpacker.compiler.send(:webpack_env)["FOO"] == nil
+    Webpacker.compiler.env["FOO"] = "BAR"
+    assert Webpacker.compiler.send(:webpack_env)["FOO"] == "BAR"
   end
 
-  def test_compile?
-    assert Webpacker::Compiler.compile?
+  def test_freshness
+    assert Webpacker.compiler.stale?
+    assert !Webpacker.compiler.fresh?
   end
 end