+++ test/manifest_test.rb
@@ -1,32 +1,36 @@
-require "webpacker_test"
+require "webpacker_test_helper"
 
 class ManifestTest < Minitest::Test
-  def test_file_path
-    file_path = File.join(File.dirname(__FILE__), "test_app/public/packs", "manifest.json").to_s
-    assert_equal Webpacker::Manifest.file_path.to_s, file_path
-  end
-
   def test_lookup_exception
-    manifest_path = File.join(File.dirname(__FILE__), "test_app/public/packs", "manifest.json").to_s
+    manifest_path = File.expand_path File.join(File.dirname(__FILE__), "test_app/public/packs", "manifest.json").to_s
     asset_file = "calendar.js"
 
-    Webpacker::Configuration.stub :compile?, false do
-      error = assert_raises Webpacker::FileLoader::NotFoundError do
-        Webpacker::Manifest.lookup(asset_file)
+    Webpacker.config.stub :compile?, false do
+      error = assert_raises Webpacker::Manifest::MissingEntryError do
+        Webpacker.manifest.lookup(asset_file)
       end
 
-      assert_equal error.message, "Can't find #{asset_file} in #{manifest_path}. Is webpack still compiling?"
+      expected = <<-MSG
+Webpacker can't find #{asset_file} in #{manifest_path}. Possible causes:
+1. You want to set wepbacker.yml value of compile to true for your environment
+   unless you are using the `webpack -w` or the webpack-dev-server.
+2. Webpack has not yet re-run to reflect updates.
+3. You have misconfigured Webpacker's config/webpacker.yml file.
+4. Your Webpack configuration is not creating a manifest.
+Your manifest contains:
+{
+  "bootstrap.css": "/packs/bootstrap-c38deda30895059837cf.css",
+  "application.css": "/packs/application-dd6b1cd38bfa093df600.css",
+  "bootstrap.js": "/packs/bootstrap-300631c4f0e0f9c865bc.js",
+  "application.js": "/packs/application-k344a6d59eef8632c9d1.js"
+}
+     MSG
+
+      assert_equal expected, error.message
     end
   end
 
   def test_lookup_success
-    asset_file = "bootstrap.js"
-    assert_equal Webpacker::Manifest.lookup(asset_file), "/packs/bootstrap-300631c4f0e0f9c865bc.js"
-  end
-
-  def test_lookup_path
-    file_path = File.join(File.dirname(__FILE__), "test_app/public/packs", "bootstrap-300631c4f0e0f9c865bc.js").to_s
-    asset_file = "bootstrap.js"
-    assert_equal Webpacker::Manifest.lookup_path(asset_file).to_s, file_path
+    assert_equal Webpacker.manifest.lookup("bootstrap.js"), "/packs/bootstrap-300631c4f0e0f9c865bc.js"
   end
 end